+++ src/main/java/dev/cheem/projects/invoicestore/dto/graphql/GraphQLInvoiceInputDTO.java
@@ -7,6 +7,6 @@ public class GraphQLInvoiceInputDTO {
 
   private String date;
   private String name;
-  private Float total;
-  private Float vatTotal;
+  private Double total;
+  private Double vatTotal;
 }
